+++ public/browser/chunk-BcDciPDq2.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";var t=[{path:``,loadComponent:()=>import(`./chunk-B7STYGZo.js`).then(o=>o.ModerationListComponent)}];export{t as MODERATION_ROUTES};