+++ public/browser/chunk-fBnwO9x12.js
@@ -1,0 +1,1 @@
+import"./chunk-CD52DvrH.js";import"./chunk-B_m5Yd57.js";import"./chunk-wEifUrYN.js";import"./chunk-DOeACeJM.js";import"./chunk-BPywBVck.js";import"./chunk-BgjWCWRg.js";import"./chunk-CTXJSbbm.js";import"./chunk-BZd4xebm.js";import"./chunk-Ck4Fw73x.js";import{M as Fe}from"./main-VH43SEGF.js";export{Fe as BpmnEditorPage};